Changed algorithm to allow stopping earlier.
If I need only to chech if a point is inside an hexa
I can stop the procedure if the target point is outside
the bounding box. However if I need to undertand where
the target point lays w.r.t. the element I need to perform
one Iteration step.